requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
python-jose
passlib[bcrypt]
python-dotenv
alembic